refactor(guest-mock-claude): split process responsibilities
